add code of conduct text and allow <br /> characters
This text has been added to encourage participation of attendees from all walks of life and to ensure
that all attendees are aware of they way they are expected to treat each other. Wroclove.rb is a public
conference and public behavior requires etiquitte which is respectful of every individual.